requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
et_xmlfile==2.0.0
fastapi==0.135.3
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
numpy==2.4.4
ollama==0.6.1
openpyxl==3.1.5
pandas==3.0.2
portalocker==3.2.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
qdrant-client==1.17.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
